% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_draft.R
\name{close_draft}
\alias{close_draft}
\title{Close an Outlook message window.}
\usage{
close_draft(ol_mail, save = FALSE)
}
\arguments{
\item{ol_mail}{a COM object that binds to an e-mail window ('Outlook.MailItem')}

\item{save}{if TRUE, save to drafts, else discard without confirmation}
}
\description{
Close an Outlook message window.
}
