% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{connect_outlook}
\alias{connect_outlook}
\title{Create a COM Outlook.Application instance.}
\usage{
connect_outlook(wait_seconds = 3)
}
\value{
An object of class `COMIDispatch.` that represents an Outlook application instance.
}
\description{
Create a COM Outlook.Application instance.
}
\examples{
\dontrun{
com <- connect_outlook()
}
}
