% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_draft.R
\name{create_draft}
\alias{create_draft}
\title{Connect to Outlook and create an e-mail message.}
\usage{
create_draft(
  ol_app,
  addr_from = NULL,
  addr_to = NULL,
  addr_cc = NULL,
  subject = NULL,
  body_html = NULL,
  body_plain = NULL,
  attachments = NULL,
  use_signature = TRUE,
  show_message = TRUE
)
}
\arguments{
\item{ol_app}{An object that represents an Outlook application instance (class: class `COMIDispatch.`)}

\item{addr_from}{optional fields to fill}

\item{addr_to}{optional fields to fill}

\item{addr_cc}{optional fields to fill}

\item{subject}{optional fields to fill}

\item{attachments}{path to file(s) to attach}

\item{use_signature}{if TRUE, get the HTML signature from the new blank message, than put it back *after the plain text* after the message is created.}

\item{show_message}{if TRUE, show the message for editing after creation}

\item{html_body}{HTML contents of the message. Takes precedence over body}

\item{body}{plain text contents of the message}
}
\value{
a COM object that binds to an e-mail window (`'Outlook.MailItem'``)
}
\description{
Connect to Outlook and create an e-mail message.
}
\examples{
\dontrun{

com <- connect_outlook()
msg <- create_draft(com, addr_to = "foo@bar.com")

# Attachments
msg <- create_draft(con, attachments = c('foo.txt', 'foo2.txt'))

}
}
