% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{disconnect_outlook}
\alias{disconnect_outlook}
\title{Close an Outlook COM instance.}
\usage{
disconnect_outlook(com)
}
\arguments{
\item{com}{An object of class `COMIDispatch.` that represents an Outlook application instance.}
}
\value{
nothing
}
\description{
Close an Outlook COM instance.
}
\examples{

disconnect_outlook(com)

}
