% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{has_COM_method}
\alias{has_COM_method}
\title{Check if a COM object has a method}
\usage{
has_COM_method(x, method_name)
}
\arguments{
\item{x}{any object}

\item{method_name}{a single method name to query, as a character}
}
\value{
TRUE if x is a COM object with the given property
}
\description{
Check if a COM object has a method
}
