% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{has_COM_property}
\alias{has_COM_property}
\title{Check if a COM object has a property (attribute)}
\usage{
has_COM_property(x, prop_name)
}
\arguments{
\item{x}{any object}

\item{prop_name}{a single property name to query, as a character}
}
\value{
TRUE if x is a COM object with the given property
}
\description{
Check if a COM object has a property (attribute)
}
