% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{is_COM}
\alias{is_COM}
\title{Check if an object is a COM object.}
\usage{
is_COM(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE or FALSE
}
\description{
Check if an object is a COM object.
}
\keyword{internal}
