% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{is_mail}
\alias{is_mail}
\title{Check if an object is an Outlook message.}
\usage{
is_mail(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE if the object is a binding to Outlook.MailItem instance.
}
\description{
Check if an object is an Outlook message.
}
