% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{is_outlook}
\alias{is_outlook}
\title{Check if an object is an Outlook application.}
\usage{
is_outlook(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE if the object is a binding to Outlook Application instance.
}
\description{
Check if an object is an Outlook application.
TRUE if the object is a COM object with the property `Name`, with value `Outlook`.
}
