% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_draft.R
\name{open_msg}
\alias{open_msg}
\title{Open a .msg in the current Outlook session}
\usage{
open_msg(ol_app, path_msg, show_message = TRUE)
}
\arguments{
\item{ol_app}{An object that represents an Outlook application instance (class: class `COMIDispatch.`)}

\item{path_msg}{a path to a '.msg' or '.eml' file (not checked for format)}

\item{show_message}{if TRUE, show the message for editing after creation}
}
\value{
a COM object that binds to an e-mail window (`'Outlook.MailItem'`)
}
\description{
Open a .msg in the current Outlook session
}
\examples{
\dontrun{

# Path to a sample .msg file
f <- system.file('data/sample.msg', package = 'outlookMailer')

con <- connect_outlook()
msg <- open_msg(con, path_msg = f, show_message = TRUE)

}
}
