% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlookMailer-package.R
\docType{package}
\name{outlookMailer-package}
\alias{outlookMailer}
\alias{outlookMailer-package}
\title{outlookMailer: Interact with Microsoft Outlook from R}
\description{
Interact with Microsoft Outlook from R.
}
\author{
\strong{Maintainer}: Lorenzo Gaborini \email{lorenzo.gaborini@ita.sport}

}
\keyword{internal}
